backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
websockets==13.1
httpx==0.28.0
aiofiles==24.1.0
boto3==1.42.64
botocore==1.42.64
awscrt
langchain==0.3.10
langchain-aws==0.2.7
langchain-community==0.3.10
faiss-cpu==1.9.0
numpy==1.26.4
Pillow==11.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.4.0
tenacity==9.0.0
aws-sdk-bedrock-runtime
smithy-core
smithy-aws-core
